requirements.txt
flask
flasgger
psycopg2-binary
ultralytics
torch
bluezero
dbus-python
numpy
pandas
scikit-learn
matplotlib
seaborn
python-dotenv
clip
shapely
lap
opencv-python
geopy
flask-cors
